+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Return `edits` in `codeAction/resolve` responses rather than `commands`. #655
   - Improve `:linters :clj-kondo :async-custom-lint?` to avoid infinite loops.
   - Add new custom LSP feature __Test Tree__, which shows all test hierarchy of a file. #653
+  - Improve function name finding to consider other function definition types for some features. #666
 
 - API/CLI
   - Exit process if any error during classpath lookup.